+++ Updates/Classes/Core/Model/NotificationMode.swift
@@ -13,4 +13,20 @@ public enum NotificationMode: String, Codable {
     case twice
     case thrice
     case always
+    case withoutAvailableUpdate = "without-available-update"
+
+    var notificationCount: Int {
+        switch self {
+        case .never:
+            return 0
+        case .once:
+            return 1
+        case .twice:
+            return 2
+        case .thrice:
+            return 3
+        case .always, .withoutAvailableUpdate:
+            return Int.max
+        }
+    }
 }