+++ Updates/Classes/Core/Updates.swift
@@ -9,20 +9,22 @@ import Foundation
 import StoreKit
 
 public class Updates {
-    
+
     // MARK: Global State
-    
+
     public static var configurationType: ConfigurationType = {
         for configurationType in ConfigurationType.allCases where bundledConfigurationURL(configurationType) != nil {
             return configurationType
         }
         return .json // Default configuration type.
     }()
-    
+
     /// Defaults configuration URL to bundled configuration and detects configuration type when set.
     public static var configurationURL: URL? = bundledConfigurationURL() {
         didSet { // detect configuration format by extension
-            guard let lastPathComponent = configurationURL?.lastPathComponent.lowercased() else { return }
+            guard let lastPathComponent = configurationURL?.lastPathComponent.lowercased() else {
+                return
+            }
             let configExtension = configurationType.rawValue.lowercased()
             for configurationType in ConfigurationType.allCases where lastPathComponent.contains(configExtension) {
                 Updates.configurationType = configurationType
@@ -30,7 +32,7 @@ public class Updates {
             }
         }
     }
-    
+
     public static var appStoreId: String? {
         didSet {
             guard appStoreURL == nil, let appStoreId = appStoreId, let productName = productName else {
@@ -39,9 +41,9 @@ public class Updates {
             appStoreURL = appStoreURL(appStoreId: appStoreId, productName: productName)
         }
     }
-    
+
     public static var appStoreURL: URL?
-    
+
     /// Returns the URL to open the app with the specified identifier in the App Store.
     /// - Parameters:
     ///     - appStoreId: The app store identifier specified as a String.
@@ -51,11 +53,49 @@ public class Updates {
         Updates.appStoreId = appStoreId
         return appStoreURL
     }
-    
+
     public static let buildString: String? = Bundle.main.infoDictionary?[kCFBundleVersionKey as String] as? String
-    
+
     public static var bundleIdentifier: String? = Bundle.main.bundleIdentifier
-    
+
+    public static var comparingVersions: VersionComparator = .patch
+
+    public static var countryCode: String? = {
+        let currentBundle = Bundle(for: Updates.self)
+        if #available(iOS 13.0, macCatalyst 13.0, *),
+            useStoreKit,
+            let iso3166Alpha3CountryCode = SKPaymentQueue.default().storefront?.countryCode,
+            !iso3166Alpha3CountryCode.isEmpty,
+            let iso3166Mapping = currentBundle.infoDictionary?["ISO3166Map"] as? [String: String],
+            let iso3166Alpha2CountryCode = iso3166Mapping[iso3166Alpha3CountryCode] {
+            return iso3166Alpha2CountryCode
+        } else {
+            return Locale.current.regionCode
+        }
+    }()
+
+    public static var newVersionString: String?
+
+    public static var notifying: NotificationMode = .once
+
+    public static var minimumOSVersion: String?
+
+    public static var minimumOptionalAppVersion: String?
+
+    public static var minimumRequiredAppVersion: String?
+
+    public static let productName: String? = Bundle.main.infoDictionary?["CFBundleDisplayName"] as? String
+
+    public static var releaseNotes: String?
+
+    public static var updateType: UpdateType = .soft
+
+    public static var updatingMode: UpdatingMode = .automatically
+
+    public static var useStoreKit = true
+
+    public static var versionString: String? = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String
+
     public static func checkForUpdates(currentOSVersion: String, completion: @escaping (UpdatesResult) -> Void) {
         guard let configURL = configurationURL, let cachedConfigURL = cachedConfigurationURL else {
             checkForUpdates(
@@ -69,7 +109,7 @@ public class Updates {
             configurationURL: configURL,
             cachedConfigurationURL: cachedConfigURL
         )
-        configurationService.fetchSettings { result in
+        configurationService.fetchSettings(defaults: programmaticConfiguration()) { result in
             switch result {
             case .success(let configuration):
                 checkForUpdates(
@@ -86,45 +126,25 @@ public class Updates {
             }
         }
     }
-    
-    private static func programmaticConfiguration() -> ConfigurationResult {
-        return ConfigurationResult(
-            appStoreId: appStoreId,
-            build: buildString,
-            comparator: comparingVersions,
-            minRequiredOSVersion: minimumOSVersion,
-            notifying: notifying,
-            releaseNotes: releaseNotes,
-            updatingMode: updatingMode,
-            version: versionString
-        )
-    }
-    
+
     private static func checkForUpdates(configuration: ConfigurationResult,
                                         operatingSystemVersion: String,
                                         completion: @escaping (UpdatesResult) -> Void) {
-        guard let bundleVersion = configuration.version, let buildString = configuration.buildString else {
-            completion(.none)
-            return
-        }
-        registerBuild(bundleVersion: bundleVersion, buildString: buildString)
         let updatesService: UpdateResolutionService
         if let bundleIdentifier = bundleIdentifier, let countryCode = countryCode,
             let appMetadataService = Services.appMetadata(
                 bundleIdentifier: bundleIdentifier,
                 countryCode: countryCode
             ) {
-            updatesService = StrategicUpdateResolutionService(
+            updatesService = Services.updateResolutionService(
                 appMetadataService: appMetadataService,
-                bundleVersion: bundleVersion,
                 configuration: configuration,
                 operatingSystemVersion: operatingSystemVersion,
                 strategy: configuration.updatingMode
             )
         } else {
-            updatesService = StrategicUpdateResolutionService(
+            updatesService = Services.updateResolutionService(
                 appMetadataService: nil,
-                bundleVersion: bundleVersion,
                 configuration: configuration,
                 operatingSystemVersion: operatingSystemVersion,
                 strategy: .manually
@@ -132,43 +152,43 @@ public class Updates {
         }
         updatesService.checkForUpdates(completion: completion)
     }
-    
-    private static func registerBuild(bundleVersion: String, buildString: String) {
-        let versionJournaling = Services.versionJournallingService()
-        _ = versionJournaling.registerBuild(versionString: bundleVersion, buildString: buildString, comparator: .build)
-    }
-    
-    public static var comparingVersions: VersionComparator = .patch
-    
-    public static var countryCode: String? = {
-        let currentBundle = Bundle(for: Updates.self)
-        if #available(iOS 13.0, macCatalyst 13.0, *),
-            let iso3166Alpha3CountryCode = SKPaymentQueue.default().storefront?.countryCode,
-            !iso3166Alpha3CountryCode.isEmpty,
-            let iso3166Mapping = currentBundle.infoDictionary?["ISO3166Map"] as? [String: String],
-            let iso3166Alpha2CountryCode = iso3166Mapping[iso3166Alpha3CountryCode] {
-            return iso3166Alpha2CountryCode
-        } else {
-            return Locale.current.regionCode
+
+    /// Warning: Use either this method or `checkForUpdates` but never both as whichever is called first will return the
+    /// correct result. If using `checkForUpdates` then an `AppUpdatedResult` is returned as part of the
+    /// `UpdatesResult` object.
+    public static func isInstallOrUpdate(
+        buildString: String? = Updates.buildString,
+        comparingVersions: VersionComparator = Updates.comparingVersions,
+        versionString: String? = Updates.versionString,
+        completion: @escaping (AppUpdatedResult) -> Void
+    ) {
+        guard let versionString = versionString else {
+            completion(.init(isFirstLaunchFollowingInstall: false, isFirstLaunchFollowingUpdate: false))
+            return
         }
-    }()
-    
-    public internal(set) static var isFirstLaunchFollowingInstall: Bool = false
-    
-    public internal(set) static var isFirstLaunchFollowingUpdate: Bool = false
-    
-    public static var newVersionString: String?
-    
-    public static var notifying: NotificationMode = .once
-    
-    public static var minimumOSVersion: String?
-    
-    public static let productName: String? = Bundle.main.infoDictionary?["CFBundleDisplayName"] as? String
-    
-    public static var releaseNotes: String?
-    
-    public static var updatingMode: UpdatingMode = .automatically
-    
-    public static var versionString: String? = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String
-    
+        let isUpdated = Services.journaling.registerBuild(
+            versionString: versionString,
+            buildString: buildString,
+            comparator: comparingVersions
+        )
+        completion(isUpdated)
+    }
+
+    private static func programmaticConfiguration() -> ConfigurationResult {
+        return ConfigurationResult(
+            appStoreId: appStoreId,
+            buildString: buildString,
+            bundleVersion: versionString,
+            comparator: comparingVersions,
+            minOptionalAppVersion: minimumOptionalAppVersion,
+            minRequiredAppVersion: minimumRequiredAppVersion,
+            minRequiredOSVersion: minimumOSVersion,
+            notifying: notifying,
+            releaseNotes: releaseNotes,
+            updateType: updateType,
+            updatingMode: updatingMode,
+            latestVersion: versionString
+        )
+    }
+
 }