+++ Updates/Classes/Services/Configuration/Implementation/DefaultConfigurationService.swift
@@ -8,28 +8,28 @@
 import Foundation
 
 struct DefaultConfigurationService: ConfigurationService {
-    
+
     private let cachedConfigurationURL: URL
     private let configurationURL: URL
     private let parsingService = ConfigurationJSONParsingService()
-    
+
     init(configurationURL: URL, cachedConfigurationURL: URL) {
         self.cachedConfigurationURL = cachedConfigurationURL
         self.configurationURL = configurationURL
     }
-    
+
     /// Asynchronously fetches confguration settings.
-    func fetchSettings(_ completion: @escaping (ConfigurationServiceResult) -> Void) {
+    func fetchSettings(defaults: ConfigurationResult, completion: @escaping (ConfigurationServiceResult) -> Void) {
         DispatchQueue.global(qos: .background).async {
-            let settings = self.fetchSettings(configurationURL: self.configurationURL)
+            let settings = self.fetchSettings(configurationURL: self.configurationURL, defaults: defaults)
             onMainQueue(completion)(settings)
         }
     }
-    
+
 }
 
 private extension DefaultConfigurationService {
-    
+
     private func cacheConfiguration(_ result: ConfigurationResult) {
         let encoder = JSONEncoder()
         guard let data = try? encoder.encode(result) else {
@@ -39,13 +39,14 @@ private extension DefaultConfigurationService {
     }
 
     /// Synchronously fetches settings from the given URL.
-    private func fetchSettings(configurationURL: URL) -> ConfigurationServiceResult {
+    private func fetchSettings(configurationURL: URL, defaults: ConfigurationResult) -> ConfigurationServiceResult {
         guard let configurationData = try? Data(contentsOf: configurationURL) else {
             return .failure(.networking)
         }
         let parsingResult = parsingService.parse(configurationData)
         switch parsingResult {
-        case .success(let configuration):
+        case .success(let result):
+            let configuration = merge(result: result, defaults: defaults)
             if configurationURL != cachedConfigurationURL {
                 cacheConfiguration(configuration)
             }
@@ -54,8 +55,26 @@ private extension DefaultConfigurationService {
             guard configurationURL != cachedConfigurationURL else {
                 return .failure(.parsing(error))
             }
-            return fetchSettings(configurationURL: cachedConfigurationURL)
+            return fetchSettings(configurationURL: cachedConfigurationURL, defaults: defaults)
         }
     }
-    
+
+    /// Fills in any information missing from Updates.json with programmatically configured default values.
+    private func merge(result: ConfigurationResult, defaults: ConfigurationResult) -> ConfigurationResult {
+        return ConfigurationResult(
+            appStoreId: result.appStoreId ?? defaults.appStoreId,
+            buildString: result.buildString ?? defaults.buildString,
+            bundleVersion: result.bundleVersion ?? defaults.bundleVersion,
+            comparator: result.comparator,
+            minOptionalAppVersion: result.minOptionalAppVersion ?? defaults.minOptionalAppVersion,
+            minRequiredAppVersion: result.minRequiredAppVersion ?? defaults.minRequiredAppVersion,
+            minRequiredOSVersion: result.minOSRequired ?? defaults.minOSRequired,
+            notifying: result.notificationMode,
+            releaseNotes: result.releaseNotes ?? defaults.releaseNotes,
+            updateType: result.updateType,
+            updatingMode: result.updatingMode,
+            latestVersion: result.latestVersion ?? defaults.latestVersion
+        )
+    }
+
 }