+++ Updates/Classes/Services/Services.swift
@@ -8,32 +8,35 @@
 import Foundation
 
 struct Services {
-    
+
     static func appMetadata(bundleIdentifier: String, countryCode: String) -> AppMetadataService? {
         return ITunesSearchAPIService(bundleIdentifier: bundleIdentifier, countryCode: countryCode)
     }
-    
+
     static func configuration(configurationURL: URL, cachedConfigurationURL: URL) -> ConfigurationService {
         return DefaultConfigurationService(
             configurationURL: configurationURL,
             cachedConfigurationURL: cachedConfigurationURL
         )
     }
-    
-    static func updateResolutionService(appMetadataService: AppMetadataService? = nil, bundleVersion: String,
-                                        configuration: ConfigurationResult, operatingSystemVersion: String,
-                                        strategy: UpdatingMode) -> UpdateResolutionService {
+
+    static var journaling: VersionJournalingService {
+        DefaultVersionJournalingService()
+    }
+
+    static func updateResolutionService(
+        appMetadataService: AppMetadataService? = nil,
+        configuration: ConfigurationResult,
+        operatingSystemVersion: String,
+        strategy: UpdatingMode
+    ) -> UpdateResolutionService {
         return StrategicUpdateResolutionService(
             appMetadataService: appMetadataService,
-            bundleVersion: bundleVersion,
             configuration: configuration,
+            journalingService: journaling,
             operatingSystemVersion: operatingSystemVersion,
             strategy: strategy
         )
     }
-    
-    static func versionJournallingService() -> VersionJournalingService {
-        return DefaultVersionJournalingService()
-    }
-    
+
 }