+++ Updates/Classes/Services/Updates/AutomaticUpdateResolutionService.swift
@@ -8,20 +8,24 @@
 import Foundation
 
 struct AutomaticUpdateResolutionService: UpdateResolutionService {
-    
+
     private let appMetadataService: AppMetadataService
-    private let bundleVersion: String
     private let configuration: ConfigurationResult
+    private let journalingService: VersionJournalingService
     private let operatingSystemVersion: String
-    
-    init(appMetadataService: AppMetadataService, bundleVersion: String,
-         configuration: ConfigurationResult, operatingSystemVersion: String) {
+
+    init(
+        appMetadataService: AppMetadataService,
+        configuration: ConfigurationResult,
+        journalingService: VersionJournalingService,
+        operatingSystemVersion: String
+    ) {
         self.appMetadataService = appMetadataService
-        self.bundleVersion = bundleVersion
         self.configuration = configuration
+        self.journalingService = journalingService
         self.operatingSystemVersion = operatingSystemVersion
     }
-    
+
     func checkForUpdates(completion: @escaping (UpdatesResult) -> Void) {
         DispatchQueue.global(qos: .background).async {
             self.appMetadataService.fetchAppMetadata { result in
@@ -30,15 +34,31 @@ struct AutomaticUpdateResolutionService: UpdateResolutionService {
                     let updatedConfiguration = self.configuration.mutableCopy(with: apiResult)
                     let factory = UpdatesResultFactory(
                         configuration: updatedConfiguration,
-                        bundleVersion: self.bundleVersion,
+                        journalingService: journalingService,
                         operatingSystemVersion: self.operatingSystemVersion
                     )
                     onMainQueue(completion)(factory.manufacture())
                 case .failure:
-                    onMainQueue(completion)(.none)
+                    guard let versionString = self.configuration.bundleVersion else {
+                        onMainQueue(completion)(
+                            .none(
+                                AppUpdatedResult(
+                                    isFirstLaunchFollowingInstall: false,
+                                    isFirstLaunchFollowingUpdate: false
+                                )
+                            )
+                        )
+                        return
+                    }
+                    let isUpdated = journalingService.registerBuild(
+                        versionString: versionString,
+                        buildString: configuration.buildString,
+                        comparator: configuration.comparator
+                    )
+                    onMainQueue(completion)(.none(isUpdated))
                 }
             }
         }
     }
-    
+
 }