+++ Updates/Classes/Services/Updates/ManualUpdateResolutionService.swift
@@ -8,26 +8,30 @@
 import Foundation
 
 struct ManualUpdateResolutionService: UpdateResolutionService {
-    
-    private let bundleVersion: String
+
     private let configuration: ConfigurationResult
+    private let journalingService: VersionJournalingService
     private let operatingSystemVersion: String
-    
-    init(configuration: ConfigurationResult, bundleVersion: String, operatingSystemVersion: String) {
-        self.bundleVersion = bundleVersion
+
+    init(
+        configuration: ConfigurationResult,
+        journalingService: VersionJournalingService,
+        operatingSystemVersion: String
+    ) {
         self.configuration = configuration
+        self.journalingService = journalingService
         self.operatingSystemVersion = operatingSystemVersion
     }
-    
+
     func checkForUpdates(completion: @escaping (UpdatesResult) -> Void) {
         DispatchQueue.global(qos: .background).async {
             let factory = UpdatesResultFactory(
                 configuration: self.configuration,
-                bundleVersion: self.bundleVersion,
+                journalingService: journalingService,
                 operatingSystemVersion: self.operatingSystemVersion
             )
             onMainQueue(completion)(factory.manufacture())
         }
     }
-    
+
 }