+++ Updates/Classes/UI/UpdatesUI.swift
@@ -32,9 +32,14 @@ public class UpdatesUI: NSObject {
     }
     
     /// Presents SKStoreProductViewController modally.
-    public func presentAppStore(animated: Bool = true, completion: (() -> Void)? = nil,
-                                presentingViewController: UIViewController) {
-        guard let appStoreId = Updates.appStoreId, let appStoreIdentifierInt = UInt(appStoreId) else {
+    public func presentAppStore(
+        animated: Bool = true,
+        appStoreId: String?,
+        appStoreURL: URL?,
+        completion: (() -> Void)? = nil,
+        presentingViewController: UIViewController
+    ) {
+        guard let appStoreId = appStoreId, let appStoreIdentifierInt = UInt(appStoreId) else {
             return
         }
         let animated = self.animated
@@ -42,12 +47,13 @@ public class UpdatesUI: NSObject {
         let parameters = [SKStoreProductParameterITunesItemIdentifier: appStoreIdentifier]
         let viewController = SKStoreProductViewController()
         viewController.delegate = self
-        viewController.loadProduct(withParameters: parameters) { [weak self] (loadedSuccessfully, error) in
-            if !loadedSuccessfully, let appStoreURL = Updates.appStoreURL {
-                print(error as Any)
-                self?.presentSafariViewController(animated: animated,
-                                                  presentingViewController: presentingViewController,
-                                                  url: appStoreURL)
+        viewController.loadProduct(withParameters: parameters) { [weak self] (loadedSuccessfully, _) in
+            if !loadedSuccessfully, let appStoreURL = appStoreURL {
+                self?.presentSafariViewController(
+                    animated: animated,
+                    presentingViewController: presentingViewController,
+                    url: appStoreURL
+                )
             }
         }
         DispatchQueue.main.async {
@@ -56,9 +62,14 @@ public class UpdatesUI: NSObject {
     }
     
     /// Presents SKStoreProductViewController modally.
-    public func presentAppStore(animated: Bool = true, delegate: SKStoreProductViewControllerDelegate,
-                                presentingViewController: UIViewController) {
-        guard let appStoreId = Updates.appStoreId, let appStoreIdentifierInt = UInt(appStoreId) else {
+    public func presentAppStore(
+        animated: Bool = true,
+        appStoreId: String?,
+        appStoreURL: URL?,
+        delegate: SKStoreProductViewControllerDelegate,
+        presentingViewController: UIViewController
+    ) {
+        guard let appStoreId = appStoreId, let appStoreIdentifierInt = UInt(appStoreId) else {
             return
         }
         let animated = self.animated
@@ -66,9 +77,8 @@ public class UpdatesUI: NSObject {
         let parameters = [SKStoreProductParameterITunesItemIdentifier: appStoreIdentifier]
         let viewController = SKStoreProductViewController()
         viewController.delegate = delegate
-        viewController.loadProduct(withParameters: parameters) { [weak self] (loadedSuccessfully, error) in
-            if !loadedSuccessfully, let appStoreURL = Updates.appStoreURL {
-                print(error as Any)
+        viewController.loadProduct(withParameters: parameters) { [weak self] (loadedSuccessfully, _) in
+            if !loadedSuccessfully, let appStoreURL = appStoreURL {
                 self?.presentSafariViewController(animated: animated,
                                                   presentingViewController: presentingViewController,
                                                   url: appStoreURL)
@@ -80,11 +90,17 @@ public class UpdatesUI: NSObject {
     }
     
     /// Prompt the user to update to the latest version
-    public static func promptToUpdate(_ result: UpdatesResult, animated: Bool = true, completion: (() -> Void)? = nil,
-                                      presentingViewController: UIViewController,
-                                      title: String? = nil,
-                                      message: String? = nil) {
-        guard case let .available(update) = result else { return }
+    public static func promptToUpdate(
+        _ result: UpdatesResult,
+        animated: Bool = true,
+        completion: (() -> Void)? = nil,
+        presentingViewController: UIViewController,
+        title: String? = nil,
+        message: String? = nil
+    ) {
+        guard case let .available(update) = result else {
+            return
+        }
         let alertTitle: String
         if let title = title {
             alertTitle = title
@@ -97,16 +113,77 @@ public class UpdatesUI: NSObject {
         let alert = UIAlertController(title: alertTitle, message: alertMessage, preferredStyle: .alert)
         let updateButtonTitle = buttonTitle("Update")
         let updateAction = UIAlertAction(title: updateButtonTitle, style: .default) { _ in
-            alert.dismiss(animated: animated, completion: completion)
-            self.presentAppStore(animated: animated, completion: completion,
-                                 presentingViewController: presentingViewController)
+            if update.updateType != .hard {
+                alert.dismiss(animated: animated, completion: completion)
+            }
+            guard let appStoreId = update.appStoreId else {
+                return
+            }
+            self.presentAppStore(
+                animated: animated,
+                appStoreId: appStoreId,
+                appStoreURL: Updates.appStoreURL(for: appStoreId),
+                completion: completion,
+                presentingViewController: presentingViewController
+            )
         }
         let cancelButtonTitle = buttonTitle("Cancel")
         let cancelAction = UIAlertAction(title: cancelButtonTitle, style: .cancel) { _ in
             alert.dismiss(animated: animated, completion: completion)
         }
         alert.addAction(updateAction)
-        alert.addAction(cancelAction)
+        if update.updateType != .hard {
+            alert.addAction(cancelAction)
+        }
+        presentingViewController.present(alert, animated: animated, completion: nil)
+    }
+    
+    /// Prompt the user to update to the latest version
+    public static func promptToUpdate(
+        newVersionString: String?,
+        appStoreId: String?,
+        animated: Bool = true,
+        completion: (() -> Void)? = nil,
+        presentingViewController: UIViewController,
+        title: String? = nil,
+        message: String,
+        updateType: UpdateType
+    ) {
+        let alertTitle: String
+        if let title = title {
+            alertTitle = title
+        } else if let productName = Updates.productName, let newVersionString = newVersionString {
+            alertTitle = "\(productName) v\(newVersionString) Available"
+        } else if let newVersionString = newVersionString {
+            alertTitle = "App Version \(newVersionString) Available"
+        } else {
+            alertTitle = "App Update Available"
+        }
+        let alert = UIAlertController(title: alertTitle, message: message, preferredStyle: .alert)
+        let updateButtonTitle = buttonTitle("Update")
+        let updateAction = UIAlertAction(title: updateButtonTitle, style: .default) { _ in
+            if updateType != .hard {
+                alert.dismiss(animated: animated, completion: completion)
+            }
+            guard let appStoreId = appStoreId else {
+                return
+            }
+            self.presentAppStore(
+                animated: animated,
+                appStoreId: appStoreId,
+                appStoreURL: Updates.appStoreURL(for: appStoreId),
+                completion: completion,
+                presentingViewController: presentingViewController
+            )
+        }
+        let cancelButtonTitle = buttonTitle("Cancel")
+        let cancelAction = UIAlertAction(title: cancelButtonTitle, style: .cancel) { _ in
+            alert.dismiss(animated: animated, completion: completion)
+        }
+        alert.addAction(updateAction)
+        if updateType != .hard {
+            alert.addAction(cancelAction)
+        }
         presentingViewController.present(alert, animated: animated, completion: nil)
     }
     
@@ -115,10 +192,18 @@ public class UpdatesUI: NSObject {
     ///     - animated: Whether or not the modal presentation is animated.
     ///     - completion: Completion closure called on SKStoreProductViewController dismissal.
     ///     - presentingViewController: View controller to present on.
-    public static func presentAppStore(animated: Bool = true, completion: (() -> Void)? = nil,
+    public static func presentAppStore(animated: Bool = true,
+                                       appStoreId: String?,
+                                       appStoreURL: URL?,
+                                       completion: (() -> Void)? = nil,
                                        presentingViewController: UIViewController) {
-        updatesUI.presentAppStore(animated: animated, completion: completion,
-                                  presentingViewController: presentingViewController)
+        updatesUI.presentAppStore(
+            animated: animated,
+            appStoreId: appStoreId,
+            appStoreURL: appStoreURL,
+            completion: completion,
+            presentingViewController: presentingViewController
+        )
     }
     
     /// Presents SKStoreProductViewController modally.
@@ -127,10 +212,17 @@ public class UpdatesUI: NSObject {
     ///     - delegate: Delegate for receiving completion delegate call if required.
     ///     - presentingViewController: View controller to present on.
     public static func presentAppStore(animated: Bool = true,
+                                       appStoreId: String?,
+                                       appStoreURL: URL?,
                                        delegate: SKStoreProductViewControllerDelegate,
                                        presentingViewController: UIViewController) {
-        updatesUI.presentAppStore(animated: animated, delegate: delegate,
-                                  presentingViewController: presentingViewController)
+        updatesUI.presentAppStore(
+            animated: animated,
+            appStoreId: appStoreId,
+            appStoreURL: appStoreURL,
+            delegate: delegate,
+            presentingViewController: presentingViewController
+        )
     }
     
 }